+++ docs/CHANGELOG.md
@@ -99,6 +99,7 @@ Version 1.2.0 (Willow): Not yet released
 - Fix `Try Avoid Players` behaviour for spawn point selection not working properly
 - Fix rare crash when rapidly switching between players in first person spectate mode
 - Fix server processing `af_ping_location_req` packets when feature was turned off
+- Fix `Solo ignore resets` trigger option not working properly
 
 [@is-this-c](https://github.com/is-this-c)
 - Fix `PgUp`, `PgDown`, `End`, and `Home` on numeric keypads